+++ CHANGES.rst
@@ -1,4 +1,4 @@
-.. py:currentmodule:: flask
+.. py:currentmodule:: collections_extended
 
 Change Log
 ==========
@@ -8,6 +8,7 @@ Version 1.1
 
 Unreleased
 
-* Added :class:`collections_extended.IndexedDict`
+* Added :class:`IndexedDict`
 * Improve efficiency for large bag operations
 * Add :meth:`setlist.swap`
+* Add :meth:`bag.count` and :class:`CountsView`